moreunit/junit/moreUnit/util/BaseToolsTest.java
package moreUnit.util;

import junit.framework.TestCase;

public class BaseToolsTest extends TestCase {

	public void testGetTestmethodNameFromMethodName() {
		String methodName = "getValue";
		assertEquals("testGetValue", BaseTools.getTestmethodNameFromMethodName(methodName));

		methodName = null;
		assertEquals("", BaseTools.getTestmethodNameFromMethodName(methodName));
		
		methodName = "";
		assertEquals("", BaseTools.getTestmethodNameFromMethodName(methodName));
	}

	public void testGetTestedClass() {
		String className = "Eins";
		assertNull(BaseTools.getTestedClass(className));
		
		className = "EinsTest";
		assertEquals("Eins", BaseTools.getTestedClass(className));
		
		className = null;
		assertNull(BaseTools.getTestedClass(className));
		
		className = "ABC";
		assertNull(BaseTools.getTestedClass(className));
	}

	public void testGetTestedMethod() {
		String methodName = "getValue";
		assertNull(BaseTools.getTestedMethod(methodName));
		
		methodName = "testGetValue";
		assertEquals("getValue", BaseTools.getTestedMethod(methodName));
		
		methodName = null;
		assertNull(BaseTools.getTestedMethod(methodName));
	}
}